Raza_Analisis/backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
passlib
pyjwt
psycopg2-binary
